_search/server/requirements.txt
pyyaml
typesense==0.10.0
beautifulsoup4
